hugging_face/model_experiments/requirements.txt
# Core ML/AI dependencies
torch>=1.7.0
torchvision>=0.6.0
transformers>=4.5.1
safetensors
huggingface_hub

# Data processing and analysis
numpy>=1.19.0
pandas>=1.3.0
scipy>=1.5.0
scikit-image>=0.17.0

# Visualization
matplotlib>=3.3.0
Pillow>=8.0.0

# Utilities
tqdm>=4.50.0
pathlib2>=2.3.0

# Optional: for video processing (if using video experiments)
imageio>=2.9.0
opencv-python>=4.5.0

# Optional: for advanced analysis
alphashape>=1.3.0
shapely>=1.4.0 